Bug fix with window geometry
The first time a user uses the program, some defualts are set for the
size of the window when it appears. There was a bug in the definition,
because QRect expects 4 ints but two ints and two floats were being
passed. This has now fixed that.

Version bump to reflect the change.

+++ src/hdf5view/mainwindow.py
@@ -281,7 +281,10 @@ class MainWindow(QMainWindow):
             self.restoreGeometry(geometry)
         else:
             geometry = self.app.desktop().availableGeometry(self)
-            self.setGeometry(QRect(0, 0, geometry.width() * 0.8, geometry.height() * 0.7))
+            self.setGeometry(QRect(0,
+                                   0,
+                                   int(geometry.width() * 0.8),
+                                   int(geometry.height() * 0.7)))
 
         # Restore the window state
         window_state = settings.value('windowState')
